fix: room update, start with only npm run start
